requirements.txt
numpy>=1.23.5          
pandas>=1.5.0           
cassandra-driver>=3.25.0  
mysql-connector-python>=8.0.0  

